+++ .gitignore
@@ -31,7 +31,7 @@ render.experimental.xml
 *.keystore
 
 # Google Services (e.g. APIs or Firebase)
-example/src/main/google-services.json
+social-feed/src/main/google-services.json
 
 # Android Profiling
 *.hprof
