--- example/build.gradle
@@ -1,47 +1,0 @@
-plugins {
-    id 'com.android.application'
-    id 'org.jetbrains.kotlin.android'
-    id 'kotlin-kapt'
-    id 'com.google.gms.google-services'
-}
-
-android {
-    namespace 'com.likeminds.feedexample'
-    compileSdk feed_versions.compile_sdk
-
-    defaultConfig {
-        applicationId "com.likeminds.feedexample"
-        minSdk feed_versions.min_sdk
-        targetSdk feed_versions.target_sdk
-        versionCode feed_versions.sdk_version_code
-        versionName "$feed_versions.sdk_version_major.$feed_versions.sdk_version_minor.$feed_versions.sdk_version_patch"
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled true
-            shrinkResources true
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-            signingConfig signingConfigs.debug
-        }
-        debug {
-            minifyEnabled false
-            shrinkResources false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-        }
-    }
-    buildFeatures {
-        dataBinding = true
-    }
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_11
-        targetCompatibility JavaVersion.VERSION_11
-    }
-    kotlinOptions {
-        jvmTarget = JavaVersion.VERSION_11
-    }
-}
-
-apply from: "lm-feed-example-dependencies.gradle"
