--- example/src/main/AndroidManifest.xml
@@ -1,56 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <application
-        android:name=".LMFeedExample"
-        android:allowBackup="false"
-        android:dataExtractionRules="@xml/data_extraction_rules"
-        android:fullBackupContent="@xml/backup_rules"
-        android:icon="@mipmap/ic_app_icon"
-        android:label="@string/app_name"
-        android:roundIcon="@mipmap/ic_app_icon_round"
-        android:supportsRtl="true"
-        android:theme="@style/Theme.Likemindsfeedandroid"
-        tools:targetApi="31">
-        <activity
-            android:name=".auth.view.AuthActivity"
-            android:exported="true"
-            android:theme="@style/LMFeedAppTheme.NoActionBar">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-
-            <intent-filter>
-                <action android:name="android.intent.action.VIEW" />
-
-                <category android:name="android.intent.category.DEFAULT" />
-                <category android:name="android.intent.category.BROWSABLE" />
-
-                <data
-                    android:host="www.examplefeed.com"
-                    android:scheme="https" />
-                <data
-                    android:host="examplefeed.com"
-                    android:scheme="https" />
-            </intent-filter>
-        </activity>
-
-        <activity
-            android:name=".MainActivity"
-            android:exported="false"
-            android:screenOrientation="portrait"
-            android:theme="@style/LMFeedAppTheme.NoActionBar" />
-
-        <service
-            android:name=".LMFeedExampleMessagingService"
-            android:exported="false">
-            <intent-filter>
-                <action android:name="com.google.firebase.MESSAGING_EVENT" />
-            </intent-filter>
-        </service>
-    </application>
-
-</manifest>
