--- example/src/main/java/com/likeminds/feedexample/universalfeed/CustomLMUniversalFeedAdminFragment.kt
@@ -1,5 +1,0 @@
-package com.likeminds.feedexample.universalfeed
-
-import com.likeminds.feed.android.core.universalfeed.view.LMFeedUniversalFeedFragment
-
-class CustomLMUniversalFeedAdminFragment : LMFeedUniversalFeedFragment()
