+++ settings.gradle
@@ -18,5 +18,5 @@ dependencyResolutionManagement {
 }
 
 rootProject.name = "likeminds-feed-android"
-include ':example'
+include ':social-feed'
 include ':likeminds-feed-android-core'
