+++ social-feed/lm-feed-social-dependencies.gradle
