+++ social-feed/proguard-rules.pro
