+++ social-feed/src/main/java/com/likeminds/feedsocial/LMFeedExampleMessagingService.kt
@@ -1,0 +1,29 @@
+package com.likeminds.feedsocial
+
+import android.util.Log
+import com.google.firebase.messaging.FirebaseMessagingService
+import com.google.firebase.messaging.RemoteMessage
+import com.likeminds.feed.android.core.LMFeedCoreApplication.Companion.LOG_TAG
+import com.likeminds.feed.android.core.pushnotification.LMFeedNotificationHandler
+
+class LMFeedExampleMessagingService : FirebaseMessagingService() {
+
+    private lateinit var mNotificationHandler: LMFeedNotificationHandler
+
+    override fun onCreate() {
+        super.onCreate()
+        mNotificationHandler = LMFeedNotificationHandler.getInstance()
+        mNotificationHandler.create(this.application)
+    }
+
+    override fun onNewToken(token: String) {
+        super.onNewToken(token)
+        Log.d(LOG_TAG, "token generated: $token")
+    }
+
+    override fun onMessageReceived(message: RemoteMessage) {
+        super.onMessageReceived(message)
+        Log.d(LOG_TAG, "message generated: ${message.data}")
+        mNotificationHandler.handleNotification(message.data)
+    }
+}
