+++ social-feed/src/main/java/com/likeminds/feedsocial/universalfeed/CustomLMUniversalFeedAdminFragment.kt
@@ -1,0 +1,5 @@
+package com.likeminds.feedsocial.universalfeed
+
+import com.likeminds.feed.android.core.universalfeed.view.LMFeedUniversalFeedFragment
+
+class CustomLMUniversalFeedAdminFragment : LMFeedUniversalFeedFragment()
