+++ social-feed/src/main/res/layout/activity_auth.xml
