+++ social-feed/src/main/res/layout/fragment_main.xml
