+++ social-feed/src/main/res/mipmap-hdpi/ic_app_icon.png
