+++ social-feed/src/main/res/values/colors.xml
