+++ social-feed/src/main/res/values/strings.xml
