+++ social-feed/src/main/res/values/styles.xml
