+++ social-feed/src/main/res/values/themes.xml
